pyproject.toml
[project]
name = "clibyai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.66.0",
    "gradio>=6.9.0",
]
